+++ packages/pluggableWidgets/datagrid-dropdown-filter-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed issue with initial filter condition.
+
 ## [2.4.2] - 2023-06-09
 
 ### Changed